(refactor): and add helpful comments
